优化占用率处理

+++ web/include/function.php
@@ -84,7 +84,7 @@ function get_used_status(){
 
 //内存占有量
     $mem_total = trim(trim($mem_info[0],'Mem: '),'k total');
-    $mem_used = trim($mem_info[1],'k used');
+    $mem_used = (int)trim($mem_info[1],'k used');
     $mem_usage = round(100*intval($mem_used)/intval($mem_total),2); //百分比
 
     return array('cpu_usage'=>$cpu_usage,'mem_usage'=>$mem_usage);
